feat: remove `__loc__` from Model and container types
BREAKING CHANGE: Previously, the Model had `__loc__` property that was
set when model was assigned to a field in a parent model. Now, this
behavior was dropped and full location will only be used if the model is
initialized or validated. Modifying fields will not show absolute
location when error is reporteds, just the location of the fragment that
was modified. Now it should be possible to reuse same model instance in
two different models if needed.