feat!: don't assign location to nested models and mutable containers
RATIONALE: I found a possible problem when writing documentation; when
field is validated inside pre- or postprocessor of outer model then
error it reports might be different depending on whether the field has
already location set or not. This also brings back the possibility of
reusal of objects and sharing across different models.